On branch v2 Your branch is up to date with 'origin/v2'.
Changes to be committed:
	modified:   app/Http/Controllers/anggotaController.php
	modified:   app/Http/Controllers/beritaController.php
	modified:   app/Http/Controllers/dokumentasiController.php
	modified:   app/Http/Controllers/eskulController.php
	modified:   app/Http/Controllers/kepsekController.php
	modified:   app/Http/Controllers/pendaftaranController.php
	modified:   app/Http/Controllers/userController.php
	modified:   app/Http/Controllers/wakasekController.php
	modified:   app/Models/User.php
	modified:   app/Models/eskul.php
	modified:   app/Models/role.php
	new file:   database/migrations/2023_07_26_023736_alter_change_column_role_in_user_table.php
	new file:   database/migrations/2023_07_26_024904_rename_column_in_table.php
	modified:   database/seeders/UserSeeder.php
	new file:   public/images/foto-berita/49ngysmRrc.png
	new file:   public/images/logo-eskul/ISXLiLDpN2.png
	modified:   resources/views/admin/anggota-eskul/anggota.blade.php
	modified:   resources/views/admin/eskul.blade.php
	modified:   resources/views/admin/pendaftaran-eskul/pendaftaran.blade.php
	modified:   resources/views/admin/user/edituser.blade.php
	modified:   resources/views/admin/user/user.blade.php
	modified:   resources/views/admin/validation/register.blade.php
	new file:   resources/views/layout/eskul/list-eskul.blade.php
	modified:   resources/views/layout/header.blade.php
	modified:   resources/views/layout/subnav/berita.blade.php
	modified:   resources/views/layout/subnav/dokumentasi.blade.php
	modified:   resources/views/layout/subnav/pagination/links.blade.php
	modified:   resources/views/layout/subnav/pendaftaran-eskul/list-eskul.blade.php
	modified:   resources/views/welcome.blade.php
	modified:   routes/web.php

+++ app/Http/Controllers/pendaftaranController.php
@@ -21,7 +21,7 @@ class pendaftaranController extends Controller
         $on = administrator::all();
 
         if ($request->has('search')) {
-            if (auth()->user()->role == 0) {
+            if (auth()->user()->role_id == 1) {
                 $pendaftaran = pendaftaran::where('nama_calon_anggota', 'like', '%' . $request->search . '%')
                     ->orWhere('id_eskul', 'like', '%' . $request->search . '%')->paginate(5);
             } else {
@@ -29,7 +29,7 @@ class pendaftaranController extends Controller
                     ->orWhere('id_eskul', 'like', '%' . $request->search . '%')->paginate(5);
             }
         } else {
-            if (auth()->user()->role == 0) {
+            if (auth()->user()->role_id == 1) {
                 $pendaftaran = pendaftaran::paginate(10);
             } else {
 
@@ -54,7 +54,7 @@ class pendaftaranController extends Controller
     }
     public function list_eskul(){
 
-        $data = eskul::all();
+        $data = eskul::paginate(12);
         // dd($detail);
         return view('layout.subnav.pendaftaran-eskul.list-eskul', compact('data')) ;
     }
@@ -96,7 +96,7 @@ class pendaftaranController extends Controller
         $pendaftar = pendaftaran::find($id);
         $eskul = eskul::where('id','=',$pendaftar->id_eskul)->first();
         $exist = anggota::where('id_eskul', '=', $pendaftar->id_eskul)->where('nis','=',$pendaftar->nis)->first();
-        if ($exist->count() > 0) {
+        if ($exist && $exist->count() > 0) {
             $pendaftar->delete();
 
             return back()->with('error','Data Sudah terdaftar diekstrakurikuler '.$eskul->nama_eskul);
@@ -151,7 +151,7 @@ class pendaftaranController extends Controller
                 $pendaftaran->where('id', '<>', $request->id);
             }
 
-            $data = [ 'valid' => ( $pendaftaran->count() == 0)  ];
+            $data = [ 'valid' => ( $pendaftaran->count() == 1)  ];
             if(!empty($pendaftaran)){
                 return $data;
             }else{
@@ -172,7 +172,7 @@ class pendaftaranController extends Controller
                 $pendaftaran->where('id', '<>', $request->id);
             }
 
-            $data = [ 'valid' => ( $pendaftaran->count() == 0)  ];
+            $data = [ 'valid' => ( $pendaftaran->count() == 1)  ];
             if(!empty($pendaftaran)){
                 return $data;
             }else{
